LPD-95818 Reorganize sxp blueprint playwright specs under 'sxp_blueprint' directory
